hibike/requirements.txt
pylint==1.7.2
pyserial==3.2.1
